uart

+++ components/sys/uart.c
@@ -1,4 +1,6 @@
 #include "uart.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/queue.h"
 
 #define BUF_SIZE (1024)
 
@@ -40,7 +42,7 @@ static bool uart_setup_port(uart_port_t port, uint8_t tx_pin, uint8_t rx_pin, in
     return true;
 }
 
-bool uart_init(uart_port_t port, uint8_t tx_pin, uint8_t rx_pin, int baudrate) {
+bool uart_init(uart_port_t port, uint8_t tx_pin, uint8_t rx_pin, uint8_t baudrate) {
     return uart_init_logical(UART_LOGICAL_GPS, port, tx_pin, rx_pin, baudrate);
 }
 
@@ -48,7 +50,7 @@ bool uart_init_logical(uart_logical_port_t logical_port,
                        uart_port_t port,
                        uint8_t tx_pin,
                        uint8_t rx_pin,
-                       int baudrate) {
+                       uint8_t baudrate) {
     if (!uart_is_valid_logical_port(logical_port)) {
         return false;
     }
