+++ components/sys/uart.h
@@ -13,14 +13,14 @@ typedef enum {
 } uart_logical_port_t;
 
 // Backward-compatible init: initializes GPS logical port.
-bool uart_init(uart_port_t port, uint8_t tx_pin, uint8_t rx_pin, int baudrate);
+bool uart_init(uart_port_t port, uint8_t tx_pin, uint8_t rx_pin, uint8_t baudrate);
 
 // New API: select logical port explicitly.
 bool uart_init_logical(uart_logical_port_t logical_port,
                        uart_port_t port,
                        uint8_t tx_pin,
                        uint8_t rx_pin,
-                       int baudrate);
+                       uint8_t baudrate);
 
 int uart_read_logical(uart_logical_port_t logical_port,
                       uint8_t *data,
